fix(cli): move lockfile check to shared utils and optimize tool result rendering
