+++ packages/cli/src/ui/components/messages/ToolResultDisplay.tsx
@@ -27,6 +27,7 @@ import { SCROLL_TO_ITEM_END } from '../shared/VirtualizedList.js';
 import { ACTIVE_SHELL_MAX_LINES } from '../../constants.js';
 import { calculateToolContentMaxLines } from '../../utils/toolLayoutUtils.js';
 import { SubagentProgressDisplay } from './SubagentProgressDisplay.js';
+import { isLockFile } from '../../utils/fileUtils.js';
 
 export interface ToolResultDisplayProps {
   resultDisplay: string | object | undefined;
@@ -41,6 +42,10 @@ export interface ToolResultDisplayProps {
 interface FileDiffResult {
   fileDiff: string;
   fileName: string;
+}
+
+function isFileDiffResult(value: unknown): value is FileDiffResult {
+  return typeof value === 'object' && value !== null && 'fileDiff' in value && 'fileName' in value;
 }
 
 export const ToolResultDisplay: React.FC<ToolResultDisplayProps> = ({
@@ -142,20 +147,12 @@ export const ToolResultDisplay: React.FC<ToolResultDisplayProps> = ({
           </Text>
         );
       }
-    } else if (
-      typeof contentData === 'object' &&
-      contentData !== null &&
-      'fileDiff' in contentData
-    ) {
+    } else if (isFileDiffResult(contentData)) {
       content = (
         <DiffRenderer
-          diffContent={
-            // eslint-disable-next-line @typescript-eslint/no-unsafe-type-assertion
-            (contentData as FileDiffResult).fileDiff
-          }
-          // eslint-disable-next-line @typescript-eslint/no-unsafe-type-assertion
-          filename={(contentData as FileDiffResult).fileName}
-          disableTruncation={true}
+          diffContent={contentData.fileDiff}
+          filename={contentData.fileName}
+          disableTruncation={!isLockFile(contentData.fileName)}
           terminalWidth={childWidth}
         />
       );
