+++ packages/cli/src/ui/utils/fileUtils.ts
@@ -17,3 +17,18 @@ export function getFileExtension(
   const ext = path.extname(filename);
   return ext ? ext.slice(1) : null;
 }
+
+/**
+ * Checks if a filename corresponds to a package lock or dependency resolution file.
+ */
+export function isLockFile(filename: string | null | undefined): boolean {
+  if (!filename) return false;
+  const lower = filename.toLowerCase();
+  return (
+    lower.endsWith('-lock.json') ||
+    lower.endsWith('.lock') ||
+    lower.endsWith('.lockb') ||
+    lower === 'pnpm-lock.yaml' ||
+    lower === 'go.sum'
+  );
+}
